requirements.txt
torch
torchvision
ipdb
trimesh
einops
hydra-core
wandb
tqdm
open3d
tabulate
zarr
multimethod
termcolor
shortuuid
diffusers
pandas
deepspeed
numpy<2
numba
easydict
timm
ftfy
chamferdist
flatdict
opencv-python
transformers
sentence_transformers
albumentations
matplotlib
pytorch-lightning